+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+- Added support for Guzzle 8 while retaining Guzzle 7 support.
+
 ## 1.1.0
 
 Changes since `1.0.0`.